requirements.txt
Django>=5.2
django-crispy-forms
crispy-bootstrap5
django-widget-tweaks
django-filter
django-debug-toolbar